Made all imports relative, moved all files related to the app inside the src folder, updated README.md to reflect these changes
