+++ src/civ7modmanager.spec
@@ -3,7 +3,7 @@
 block_cipher = None
 
 a = Analysis(
-    ['src/main.py'],
+    ['main.py'],
     pathex=[],
     binaries=[('lib/UnRAR.exe', 'lib')],
     datas=[],
